backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
pydantic==2.11.0
pydantic-settings==2.13.1
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.27.2
passlib==1.7.4
alembic==1.13.1
aiosqlite==0.21.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.3.0
sendgrid==6.11.0
resend==0.8.0
pywebpush==2.0.0
cryptography==42.0.0
